:recycle: Remove dependency of ContextSearch on SnapShot
